third_party/WebKit/LayoutTests/http/tests/devtools/profiler/heap-snapshot-statistics-expected.txt

CONSOLE WARNING: line 3078: Element.createShadowRoot is deprecated and will be removed in M73, around March 2019. Please use Element.attachShadow instead. See https://www.chromestatus.com/features/4507242028072960 for more details.
Tests Statistics view of detailed heap snapshots.

Profiler was enabled.
Detailed heap profiles were enabled.

Running: testStatistics
hidden node size: 1
array node size: 20
string node size: 300
object node size: 4000
code node size: 50000
closure node size: 600000
regexp node size: 7000000
native node size: 80000000
bigint node size: 900000000
{"total":1887654341,"v8heap":1807654341,"native":80000000,"code":50000,"jsArrays":20,"strings":300,"system":900000000}

Profiler was disabled.
